Move client-opts to a separate file
(cherry picked from commit 01eb35bfb3839436c2e5d2182b5f200008fd58a8)
